requirements.txt
annotated-types==0.5.0
asgiref==3.7.2
crispy-bootstrap4==2022.1
dj-database-url==2.1.0
Django==4.2.5
django-bootstrap-datepicker-plus==5.0.4
django-crispy-forms==2.0
django-filter==23.3
mysql==0.0.3
mysql-connector-python==8.1.0
mysqlclient==2.2.0
Pillow==10.0.1
protobuf==4.21.12
pydantic==2.3.0
pydantic_core==2.6.3
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
